[refacto] #169 match the previous refactor
