+++ src/main/java/com/dinosaur/dinosaurexploder/constants/EntityType.java
@@ -4,5 +4,5 @@ package com.dinosaur.dinosaurexploder.constants;
  *  This handles with all the entities in the game like lives, player projectile etc.
  */
 public enum EntityType {
-    PLAYER, GREEN_DINO, PROJECTILE, ENEMY_PROJECTILE, SCORE, LIFE, BOMB, COIN, LEVEL
+    PLAYER, GREEN_DINO, PROJECTILE, ENEMY_PROJECTILE, SCORE, LIFE, BOMB, COIN, HEART, LEVEL
 }
